% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ectrl_pub.R
\name{copy_bookdown_template_files}
\alias{copy_bookdown_template_files}
\title{Create Bookdown Lite Project Template}
\usage{
copy_bookdown_template_files(path)
}
\description{
Code was copied from the \href{https://github.com/rstudio/bookdown/blob/master/R/skeleton.R}{bookdwon package}.
}
