% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ectrl_pub.R
\name{create_ectrl_pub}
\alias{create_ectrl_pub}
\title{Create a new directory with the specified publication template}
\usage{
create_ectrl_pub(dirname = "new-dir", template = "ectrl-publication")
}
\arguments{
\item{dirname}{name of the directory to create. If not specified, a directory
\emph{new-dir} will be created.}

\item{template}{which template to use?}
}
\description{
This function creates a new directory under the current directory, which will
contain the given files of the specified template, ready to be rendered.
}
\details{
The available templates are: \emph{eurocontrol-publication} (default).
(Others could be added later.)
The name of the main \code{.Rmd} file will be \emph{index.Rmd}.
}
\examples{
\dontrun{
create_ectrl_pub(dirname = "tmp_dir")
}

}
